requirements.txt
google-genai>=0.3.0
requests
python-dotenv
ffmpeg-python
PyJWT
pydantic
